Add initial set of CI tests using Standard Test Interface
Adding initial set of basic tests for qemu according to the
Standard Test Interface [1]. These tests are run using the Avocado
framework [2] using the Avocado-VT plugin [3].

The key here is to have nested virtualization enabled on the VM
that will run these tests. This patchset has been tested by me on
a Fedora 28 and Rawhide VMs.

Without nested virtualization support, the libvirtd service fails to start
or Avocado fails as the kvm_module required for the tests to run is
not loaded.

[1] https://fedoraproject.org/wiki/CI/Standard_Test_Interface
[2] https://avocado-framework.readthedocs.io/en/63.0/Introduction.html
[3] http://avocado-vt.readthedocs.io/en/latest/

Signed-off-by: Yash Mankad <ymankad@redhat.com>